pyproject.toml
[project]
name = "repo-goalie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.2",
    "python-dotenv>=1.1.1",
    "typer>=0.16.0",
]

[project.scripts]
goalie = "goalie:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["goalie"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
